git-svn-id: https://xrealengine.svn.codeplex.com/svn@43811 142e9c35-602f-4880-9f37-ec48a24daa6a

+++ trunk/XRealEngine/XRealEngine/Graphics/MapLayer.cs
@@ -30,7 +30,7 @@ namespace XRealEngine.Framework.Graphics
             set { scale = value; }
         }
 
-        public MapLayer(int scale, Color color, int depth)
+        public MapLayer(int scale, Color color, float depth)
         {
             this.Color = color;
             this.Scale = scale;
